Fix issue with missing frames not on root link
Before this commit, frames that were present on a subassembly that is
not the main trunk assembly (first in the list in Onshape) would not be
added to the URDF. This was occurring because such a frame was being
associated to the parent subassembly as read in from Onshape, BUT this
fails to account for the reassignment process for non-trunk parts.

Refactor the functionality for appending frames to improve readability
and reduce duplicate code. Fix the bug by looking up the proper parent
assignment. Add print statements to improve diagnosability.