Fix suppression and mate groups
Previously, an occurrence would only be suppressed if it was directly
marked as suppressed. Change the logic to also check parent assemblies
and subassemblies and propagate suppression down to the occurrence.

Fix an issue looking up occurrences in the assignParts method.

Improve getOccurrence function so it can look up a part by ID if look-up
by path fails for some reason.